+++ packages/amplify-dynamodb-simulator/CHANGELOG.md
@@ -3,13 +3,13 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [1.13.1](https://github.com/aws-amplify/amplify-cli/compare/amplify-dynamodb-simulator@1.13.0...amplify-dynamodb-simulator@1.13.1) (2020-01-24)
+## [1.13.2](https://github.com/aws-amplify/amplify-cli/compare/amplify-dynamodb-simulator@1.13.1...amplify-dynamodb-simulator@1.13.2) (2020-02-07)
 
 **Note:** Version bump only for package amplify-dynamodb-simulator
 
+## [1.13.1](https://github.com/aws-amplify/amplify-cli/compare/amplify-dynamodb-simulator@1.13.0...amplify-dynamodb-simulator@1.13.1) (2020-01-24)
 
-
-
+**Note:** Version bump only for package amplify-dynamodb-simulator
 
 # [1.13.0](https://github.com/aws-amplify/amplify-cli/compare/amplify-dynamodb-simulator@0.23.0...amplify-dynamodb-simulator@1.13.0) (2020-01-23)