+++ packages/amplify-graphiql-explorer/CHANGELOG.md
@@ -3,13 +3,13 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [1.1.9](https://github.com/aws-amplify/amplify-cli/compare/amplify-graphiql-explorer@1.1.8...amplify-graphiql-explorer@1.1.9) (2020-01-24)
+## [1.1.10](https://github.com/aws-amplify/amplify-cli/compare/amplify-graphiql-explorer@1.1.9...amplify-graphiql-explorer@1.1.10) (2020-02-07)
 
 **Note:** Version bump only for package amplify-graphiql-explorer
 
+## [1.1.9](https://github.com/aws-amplify/amplify-cli/compare/amplify-graphiql-explorer@1.1.8...amplify-graphiql-explorer@1.1.9) (2020-01-24)
 
-
-
+**Note:** Version bump only for package amplify-graphiql-explorer
 
 ## [1.1.8](https://github.com/aws-amplify/amplify-cli/compare/amplify-graphiql-explorer@0.23.0...amplify-graphiql-explorer@1.1.8) (2020-01-23)