+++ packages/graphql-transformer-common/CHANGELOG.md
@@ -3,13 +3,13 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [4.13.1](https://github.com/aws-amplify/amplify-cli/compare/graphql-transformer-common@4.13.0...graphql-transformer-common@4.13.1) (2020-01-24)
+## [4.13.2](https://github.com/aws-amplify/amplify-cli/compare/graphql-transformer-common@4.13.1...graphql-transformer-common@4.13.2) (2020-02-07)
 
 **Note:** Version bump only for package graphql-transformer-common
 
+## [4.13.1](https://github.com/aws-amplify/amplify-cli/compare/graphql-transformer-common@4.13.0...graphql-transformer-common@4.13.1) (2020-01-24)
 
-
-
+**Note:** Version bump only for package graphql-transformer-common
 
 # [4.13.0](https://github.com/aws-amplify/amplify-cli/compare/graphql-transformer-common@3.32.0...graphql-transformer-common@4.13.0) (2020-01-23)